+++ R/ext_biomass_browse.R
@@ -614,6 +614,13 @@ BiomassBrowse <- R6Class(
       } else {
         private$.LogFile <- .relPath(value, self$path)
       }
+    },
+
+    output_files = function(value) {
+      if (!missing(value)) {
+        stop("`output_files` is read-only", call. = FALSE)
+      }
+      c(self$LogFile)
     }
   )
 )