+++ R/ext_biomass_harvest.R
@@ -201,6 +201,13 @@ BiomassHarvest <- R6Class(
       } else {
         private$.SummaryLog <- .relPath(value, self$path)
       }
+    },
+
+    output_files = function(value) {
+      if (!missing(value)) {
+        stop("`output_files` is read-only", call. = FALSE)
+      }
+      c(self$EventLog, self$SummaryLog)
     }
   )
 )