+++ R/ext_climate_bda.R
@@ -184,6 +184,13 @@ ClimateBDA <- R6Class(
       } else {
         private$.LogFile <- .relPath(value, self$path)
       }
+    },
+
+    output_files = function(value) {
+      if (!missing(value)) {
+        stop("`output_files` is read-only", call. = FALSE)
+      }
+      c(self$LogFile)
     }
   )
 )