+++ R/ext_eda.R
@@ -158,6 +158,13 @@ EDA <- R6Class(
       } else {
         private$.LogFile <- .relPath(value, self$path)
       }
+    },
+
+    output_files = function(value) {
+      if (!missing(value)) {
+        stop("`output_files` is read-only", call. = FALSE)
+      }
+      c(self$LogFile)
     }
   )
 )