+++ R/ext_linear_wind.R
@@ -410,6 +410,13 @@ LinearWind <- R6Class(
       } else {
         private$.LogFile <- .relPath(value, self$path)
       }
+    },
+
+    output_files = function(value) {
+      if (!missing(value)) {
+        stop("`output_files` is read-only", call. = FALSE)
+      }
+      c(self$LogFile)
     }
   )
 )