+++ R/ext_original_wind.R
@@ -166,6 +166,13 @@ OriginalWind <- R6Class(
       } else {
         private$.EventLogFile <- .relPath(value, self$path)
       }
+    },
+
+    output_files = function(value) {
+      if (!missing(value)) {
+        stop("`output_files` is read-only", call. = FALSE)
+      }
+      c(self$SummaryLogFile, self$EventLogFile)
     }
   )
 )