+++ R/ext_output_landscape_habitat.R
@@ -404,6 +404,13 @@ OutputLandscapeHabitat <- R6Class(
       } else {
         private$.LogFile <- .relPath(value, self$path)
       }
+    },
+
+    output_files = function(value) {
+      if (!missing(value)) {
+        stop("`output_files` is read-only", call. = FALSE)
+      }
+      c(self$LogFile)
     }
   )
 )