+++ R/ext_root_rot.R
@@ -423,6 +423,13 @@ RootRot <- R6Class(
       } else {
         private$.SummaryLog <- .relPath(value, self$path)
       }
+    },
+
+    output_files = function(value) {
+      if (!missing(value)) {
+        stop("`output_files` is read-only", call. = FALSE)
+      }
+      c(self$EventLog, self$SummaryLog) ## NULLs dropped automatically
     }
   )
 )